policies-role-permission and event order created

+++ app/Http/Controllers/ContactUsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\ContactUs;
 use App\Models\Offer;
+use Illuminate\Support\Facades\Gate;
 
 class ContactUsController extends Controller
 {
@@ -12,6 +13,10 @@ class ContactUsController extends Controller
      */
     public function index()
     {
+        Gate::define('list contacts',function($user){
+            return $user->hasPermissionTo('list contacts');
+        });
+        Gate::authorize('list contacts');
         $contacts = ContactUs::when(request()->has('search'), function ($query) {
             $filteredColumns = ['name', 'email', 'phone', 'message','type','created_at'];
             $query->where(function ($query) use ($filteredColumns) {
@@ -29,6 +34,10 @@ class ContactUsController extends Controller
     public
     function destroy(Offer $offer)
     {
+        Gate::define('delete contact',function($user){
+            return $user->hasPermissionTo('delete contact');
+        });
+        Gate::authorize('delete contact');
         $offer->delete();
         return redirect()->route('offers.index')->with('success', "Offer for " . $offer->seller->restaurant_name . " deleted successfully");
     }
