+++ app/Http/Controllers/OfferController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Offer;
 use App\Models\Seller;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\Rule;
 
 class OfferController extends Controller
@@ -14,6 +15,11 @@ class OfferController extends Controller
      */
     public function index()
     {
+        Gate::define('list offers',function($user){
+            return $user->hasPermissionTo('list offers');
+        });
+        Gate::authorize('list offers');
+
         $offers = Offer::when(request()->has('search'), function ($query) {
             $filteredColumns = ['name', 'brief_description', 'start_date', 'end_date'];
             $query->where(function ($query) use ($filteredColumns) {
@@ -31,6 +37,10 @@ class OfferController extends Controller
      */
     public function destroy(Offer $offer)
     {
+        Gate::define('delete offer',function($user){
+            return $user->hasPermissionTo('delete offer');
+        });
+        Gate::authorize('delete offers');
         $offer->delete();
         return redirect()->route('offers.index')->with('success', "Offer for " . $offer->seller->restaurant_name . " deleted successfully");
     }
